packages/speaker_id/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ququ-speaker-id-tool"
version = "0.1.0"
description = "Spectrum-clustering baseline for host/guest speaker ID on Ququ call parts."
requires-python = ">=3.12"
dependencies = [
  "librosa>=0.11.0",
  "numpy>=1.26,<2",
  "PyYAML>=6.0.0",
  "requests>=2.31.0",
  "rich>=14.0.0",
  "scikit-learn>=1.8.0",
  "scipy>=1.17.0",
  "soundfile>=0.13.0",
  "tqdm>=4.67.0",
  "typer>=0.24.0",
]

[project.optional-dependencies]
wespeaker = [
  "torch>=2.11.0",
  "torchaudio>=2.11.0",
]

[project.scripts]
ququ-speaker-id = "speaker_id_tool.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["speaker_id_tool*"]
